Mass items performance improved in monitor mode
When the UI thread detects it's been blocked for a short while, busy inserting new items, it signals the read thread to copy everything into its local list and give back the complete list when the readers wait the next time. Timeouts and queue lengths have been decreased. All items remain visible and user interaction is possible (mostly). The counter indicator shows the total number of log items counted in the read thread and indicates something big coming... On the read waiting event (OnReadWaiting method), the ObservableCollection is replaced again, but keeping the selection/updating the scroll state when possible.