ASP.NET applications know their path and file name
The ASP.NET build detects the web application's path when calling one of the log path setup methods, so that anything that required Assembly.GetEntryAssembly() can now use the information from other sources of the AppDomain. The default log directory is not in the bin directory, because any change in the bin directory would immediately unload the application to restart it for the next request. Instead, the log directory is one level up. The FieldLog configuration file, now named web.flconfig, is also searched in the application base directory, not in bin. ASP.NET applications can now use the simple AcceptLogFileBasePath() method and don't need to specify the log directory and prefix anymore.