Fixed window flashing on new items
The FieldLogViewer window would not flash anymore on new items if it was activated while no more items arrived. This state is now correctly reset on window activation so it can flash again on the next item.

Also, the flashing now immediately stops when the window is activated.

Created the batch activity test in the ConsoleDemo application to have some time between log items.