Fixed list item focus on new items
When new items were added to the list and the CollectionView.Refresh method was called, the entire list of log items was rebuilt and didn't keep the focused item focused. When this happened while navigating through items with the keyboard, the selection was easily reset to the first list item. Instead of refreshing the entire view, we now determine which items were modified with new data and need to be updated. This is done for each item separately so that the items all remain in the list.