Always reuse files older than a day as long as they're smaller than 20 KiB
If a long-running application writes only few log items, numerous files of a few hundred bytes are created each day just because they're older than a day. This limit was created to keep the purge granularity low. But log files so small don't matter when submitting them, so they can be used again, reducing the number of files.