Fixed unhandled ArgumentException at Path.GetFileNameWithoutExtension
A user has reported an unhandled ArgumentException("Illegal characters in path.") when getting the filename from the module path in a stack frame. No further information is available. When the exception occurs, the offending path and the exception is logged, and the entire module path is used instead. This may also affect deobfuscation, where the module will likely not be found if not processed.