Don't write log files until something relevant happens
No log files are created or written to as long as only LogStart, LogShutdown, or first-chance exception items should be written. Log file writing is not started and buffers are kept alive until a relevant log item appears. In that case, the buffer (that contains irrelevant messages) is written to files normally. Once a relevant item was logged, everything is written immediately as before.